Allow visitConsentDemo to hang out with its friends
